stt-module/requirements.txt
faster-whisper>=0.9.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
minio==7.2.15
pika==1.3.2

